+++ src/api/capsule/fetch-capules.ts
@@ -7,9 +7,8 @@ type FetchCapsulesResponse = {
 };
 
 export const fetchCapsules = async (
-  props: AxiosRequestConfig
+  props?: AxiosRequestConfig
 ): Promise<FetchCapsulesResponse> => {
   const response = await selenAPIClient.get("/capsules", props);
-
   return response.data;
 };