+++ src/features/user/hooks/use-fetch-me.ts
@@ -1,10 +1,11 @@
 import { fetchMe, FetchMeResponse } from "@/src/api/user/fetch-me";
 import { useQuery } from "@tanstack/react-query";
 
-export const useFetchMe = (token: string | null) => {
+export const useFetchMe = (bearerTokenSelen: string | null) => {
   return useQuery<FetchMeResponse>({
     queryKey: ["me"],
-    queryFn: () => fetchMe({ headers: { Authorization: `Bearer ${token}` } }),
-    enabled: !!token,
+    queryFn: () =>
+      fetchMe({ headers: { Authorization: `Bearer ${bearerTokenSelen}` } }),
+    enabled: !!bearerTokenSelen,
   });
 };